pyproject.toml
[project]
name = "python-discord-scheduler-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohappyeyeballs==2.4.0",
    "aiohttp==3.10.5",
    "aiosignal==1.3.1",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "astroid==3.2.4",
    "async-timeout==4.0.3",
    "attrs==24.2.0",
    "beautifulsoup4==4.12.3",
    "black==24.8.0",
    "cachetools==5.5.2",
    "certifi==2024.8.30",
    "cffi==1.17.1",
    "charset-normalizer==3.3.2",
    "click==8.1.7",
    "cloudscraper==1.2.71",
    "codecov==2.1.13",
    "coverage==7.6.1",
    "cycler==0.12.1",
    "dill==0.3.8",
    "discord-py==2.4.0",
    "distro==1.9.0",
    "exceptiongroup==1.2.2",
    "fonttools==4.53.1",
    "frozenlist==1.4.1",
    "google-ai-generativelanguage==0.6.15",
    "google-api-core==2.25.1",
    "google-api-python-client==2.178.0",
    "google-auth==2.39.0",
    "google-auth-httplib2==0.2.0",
    "google-genai==1.11.0",
    "google-generativeai==0.8.5",
    "googleapis-common-protos==1.70.0",
    "grpcio==1.74.0",
    "grpcio-status==1.71.2",
    "gtts==2.5.3",
    "h11==0.14.0",
    "httpcore==1.0.8",
    "httplib2==0.22.0",
    "httpx==0.28.1",
    "idna==3.8",
    "importlib-resources==6.4.5",
    "iniconfig==2.0.0",
    "isort==5.13.2",
    "jiter==0.10.0",
    "kiwisolver==1.4.7",
    "matplotlib==3.9.2",
    "mccabe==0.7.0",
    "multidict==6.0.5",
    "mypy==1.15.0",
    "mypy-extensions==1.0.0",
    "networkx==3.2.1",
    "numpy==2.0.2",
    "openai==1.97.0",
    "outcome==1.3.0.post0",
    "packaging==24.1",
    "pandas==2.2.3",
    "pandas-stubs==2.2.2.240807",
    "pathspec==0.12.1",
    "pillow==10.4.0",
    "platformdirs==4.3.3",
    "plotly==5.24.1",
    "pluggy==1.5.0",
    "proto-plus==1.26.1",
    "protobuf==5.29.5",
    "psutil==7.0.0",
    "py-spy==0.4.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycodestyle==2.12.1",
    "pycparser==2.22",
    "pydantic==2.11.3",
    "pydantic-core==2.33.1",
    "pylint==3.2.7",
    "pynacl==1.5.0",
    "pyparsing==3.1.4",
    "pysocks==1.7.1",
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "pytest-cov==6.0.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "python-louvain==0.16",
    "pytz==2024.1",
    "pyvirtualdisplay==3.0",
    "requests==2.32.3",
    "requests-toolbelt==1.0.0",
    "rsa==4.9.1",
    "seaborn==0.13.2",
    "selenium==4.26.1",
    "simple-term-menu==1.6.4",
    "six==1.16.0",
    "sniffio==1.3.1",
    "sortedcontainers==2.4.0",
    "soupsieve==2.6",
    "table2ascii==1.1.3",
    "tenacity==9.0.0",
    "tomli==2.0.1",
    "tomlkit==0.13.2",
    "tqdm==4.67.1",
    "trio==0.27.0",
    "trio-websocket==0.11.1",
    "types-beautifulsoup4==4.12.0.20250204",
    "types-cachetools==6.0.0.20250525",
    "types-html5lib==1.1.11.20241018",
    "types-networkx==3.4.2.20241227",
    "types-psutil>=7.2.0.20251228",
    "types-python-dateutil==2.9.0.20241206",
    "types-pytz==2025.1.0.20250204",
    "types-requests==2.32.0.20241016",
    "types-seaborn==0.13.2.20250111",
    "typing-extensions==4.12.2",
    "typing-inspection==0.4.0",
    "tzdata==2024.2",
    "tzlocal==5.2",
    "undetected-chromedriver==3.5.5",
    "uritemplate==4.2.0",
    "urllib3==2.2.3",
    "wcwidth==0.2.13",
    "webdriver-manager==4.0.2",
    "websocket-client==1.8.0",
    "websockets==13.1",
    "wsproto==1.2.0",
    "xvfbwrapper==0.2.9",
    "yarl==1.9.6",
    "zipp==3.20.2",
]
[tool.black]
line-length = 120


[[tool.mypy.overrides]]
module = ["google.genai.*"]
follow_untyped_imports = true
